NIFI-830 move InvokeHTTP and InvokeHTTPTest back to the standard package in order to maintain compatibility for upgrades
