+++ minifi/minifi-bootstrap/src/test/resources/InvokeHttpMiNiFiProxyPasswordTemplateTest.yml
@@ -105,7 +105,7 @@ Processors:
     Routing Strategy: Route to Property name
 - id: 66649998-3b08-4b41-9f07-a51d999743f6
   name: Search Google
-  class: org.apache.nifi.processors.standard.http.InvokeHTTP
+  class: org.apache.nifi.processors.standard.InvokeHTTP
   max concurrent tasks: 1
   scheduling strategy: TIMER_DRIVEN
   scheduling period: 0 sec
