+++ minifi/minifi-toolkit/minifi-toolkit-configuration/src/test/resources/InvokeHttpMiNiFiTemplateTest-v2.yml
@@ -104,7 +104,7 @@ Processors:
     Routing Strategy: Route to Property name
 - name: Search Google
   id: 66649998-3b08-4b41-9f07-a51d999743f6
-  class: org.apache.nifi.processors.standard.http.InvokeHTTP
+  class: org.apache.nifi.processors.standard.InvokeHTTP
   max concurrent tasks: 1
   scheduling strategy: TIMER_DRIVEN
   scheduling period: 0 sec
