+++ minifi/minifi-toolkit/minifi-toolkit-configuration/src/test/resources/InvokeHttpMiNiFiTemplateTest.xml
@@ -617,7 +617,7 @@ limitations under the License.
         <name>Retry</name>
       </relationships>
       <style/>
-      <type>org.apache.nifi.processors.standard.http.InvokeHTTP</type>
+      <type>org.apache.nifi.processors.standard.InvokeHTTP</type>
     </processors>
     <remoteProcessGroups>
       <id>d41d8cd9-8f00-3204-a980-0998ecf8427e</id>
