+++ minifi/minifi-toolkit/minifi-toolkit-configuration/src/test/resources/MINIFI-496/dual_rpgs.xml
@@ -1028,7 +1028,7 @@ limitations under the License.
             </relationships>
             <state>STOPPED</state>
             <style/>
-            <type>org.apache.nifi.processors.standard.http.InvokeHTTP</type>
+            <type>org.apache.nifi.processors.standard.InvokeHTTP</type>
         </processors>
         <processors>
             <id>99d947c2-75ae-339f-0000-000000000000</id>
